added map to render admin users in in table
